pyproject.toml
[project]
name = "crypto-quant-web"
version = "0.1.0"
description = "crypto小工具"
authors = [
    {name = "mossloo", email = "wzybmw888@163.com"},
]
dependencies = [
    "Django==4.2",
    "pandas>=2.0.1",
    "drf-yasg>=1.21.5",
    "ccxt>=3.0.80",
    "djangorestframework>=3.14.0",
    "django-filter>=23.1",
    "djoser>=2.2.0",
    "pillow>=9.5.0",
    "djangorestframework-simplejwt>=5.2.2",
    "django-cors-headers>=3.14.0",
    "celery>=5.2.7",
    "redis>=4.5.4",
    "django-allauth>=0.54.0",
    "mysqlclient>=2.1.1",
    "django-imagekit>=4.1.0",
    "drf-extensions>=0.7.1",
    "django-redis>=5.2.0",
    "django-redis-cache>=1.8.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

